Final_Project/requirements.txt
absl-py==1.2.0
altair==4.2.0
apache-beam==2.40.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==20.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
boto3==1.24.48
botocore==1.27.48
bravado==11.0.3
bravado-core==5.17.0
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==7.1.2
cloudpickle==2.1.0
commonmark==0.9.1
crcmod==1.7
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
docker==4.4.4
docopt==0.6.2
entrypoints==0.4
fastavro==1.5.4
fasteners==0.17.3
fastjsonschema==2.16.1
flatbuffers==1.12
fonttools==4.34.4
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==1.32.0
google-api-python-client==1.12.11
google-apitools==0.5.31
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.16.1
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.14.1
google-cloud-bigtable==1.7.2
google-cloud-core==1.7.3
google-cloud-datastore==1.15.5
google-cloud-dlp==3.8.0
google-cloud-language==1.3.2
google-cloud-pubsub==2.13.4
google-cloud-pubsublite==1.4.2
google-cloud-recommendations-ai==0.2.0
google-cloud-resource-manager==1.6.0
google-cloud-spanner==1.19.3
google-cloud-storage==2.2.1
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-gcp==0.2.2
grpcio-status==1.47.0
h5py==3.7.0
hdfs==2.7.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==0.14.1
jsonpointer==2.3
jsonref==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keras==2.9.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kiwisolver==1.4.4
kt-legacy==1.0.4
kubernetes==12.0.1
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
ml-metadata==1.9.0
ml-pipelines-sdk==1.9.1
monotonic==1.6
msgpack==1.0.4
nbclient==0.6.6
nbconvert==6.5.1
nbformat==5.4.0
neptune-client==0.16.5
neptune-tensorflow-keras==0.9.9
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.0
opt-einsum==3.3.0
orjson==3.7.11
overrides==6.2.0
packaging==20.9
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
plotly==5.9.0
portpicker==1.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
pydot==1.4.2
pyfarmhash==0.3.2
Pygments==2.12.0
PyJWT==2.4.0
pymongo==3.12.3
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3987==1.3.8
rich==12.5.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
streamlit==1.11.1
strict-rfc3339==0.7
swagger-spec-validator==2.7.4
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-data-validation==1.9.0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.9.0
tensorflow-model-analysis==0.40.0
tensorflow-serving-api==2.9.1
tensorflow-transform==1.9.0
termcolor==1.1.0
terminado==0.15.0
tfx==1.9.1
tfx-bsl==1.9.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uritemplate==3.0.1
urllib3==1.26.11
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.2
widgetsnbextension==3.6.1
wrapt==1.14.1
xgboost==1.6.1
zipp==3.8.1
